.github/skills/repo-memory/requirements.txt
azure-cosmos>=4.7.0
azure-identity>=1.15.0
pydantic>=2.0
